[FIX] forecast_line: issue in multicompany
when running in multi company, sometimes the user making a change is not connected
using the company of the record. In this case the default value for company_id on
forecast.line will lead to a wrong forecast.

we fix this by using the best company when updating the forecast (the
employee's company for task assigned to employees or leaves, the sale
line's company for sales forecast